Fix limited character-range
This proposal tries to fix the currently limited character range to `char`. This didn't work without the following rules:

- A `name` always end with a closing bracket (`]`) and an `eol`.
- A `key` never begins with an opening bracket (`[`) and it never contains an equal-sign (`=`).

To avoid these rules i see only the option to introduce escaping. @jedmao  - i am really curious what you think about it?

---

I also restructured it a little bit to meet the common practice having atomic and tokens at the end of the grammar. I feel like the order of the current rules is a bit more readable if you're not used to read this grammar. But it's only a suggestion.